v0.2.0: Python 3 / NumPy 2.x compatibility fixes and test suite
All library changes fix Python 2 relics or NumPy/SciPy API breakage on
the current Python 3.13 + NumPy 2.x + SciPy stack:

- EMMap.py, EMMap_noheadoverwrite.py: .tostring() → .tobytes(); str+bytes
  concatenation in MRC label; resample num float → int(round(...)) for
  scipy.signal.resample
- symm_checker.py: file() → open(); integer division /→//; Fourier subplot
  slice clamped to avoid out-of-bounds on large step values
- PEET2Pytom.py, PEETPRMParser.py, relion_dose_file_maker.py,
  star_to_newstack.py: file() → open() / savetxt(file()) → savetxt()
- PEETPicker.py: operator precedence fix (diameter/2. applied after unit())
- PEETParticleAnalysis.py: ragged array created with dtype=object
- PEETParticleCleanup.transfer_offsets_to_model: early-exit guard for
  empty particle lists (was crashing with IndexError)
- conversions.py, make_shapes.py: minor import and int-cast fixes

Adds a pytest suite covering 83 test cases (6 skipped) across
conversions, model parser, motive list, PRM parser, and all pex commands.
Test data lives in not_to_be_distributed_test_data/ (gitignored).

+++ TEMPy_extensions_py3/make_shapes.py
@@ -31,8 +31,8 @@ def make_cylinder(box_size, radius, height, symm_axis=1, centre=False, apix=1, f
     new_map = fill*(new_map<=radius**2)
     new_map = Map(new_map, origin, apix, 'cyl')
 
-    botcut = new_map.pixel_centre()[symm_axis] + height/2
-    topcut = new_map.pixel_centre()[symm_axis] - height/2
+    botcut = int(new_map.pixel_centre()[symm_axis] + height/2)
+    topcut = int(new_map.pixel_centre()[symm_axis] - height/2)
 
     if symm_axis == 0:
         new_map.fullMap[:,:,:topcut] = np.zeros([box_size[2], box_size[1], topcut])
